feat: extract a common interface for TraditionViewContent
