+++ demo/src/main/java/com/cardinalblue/kraftshade/demo/ui/view/TraditionViewContent.kt
@@ -1,0 +1,8 @@
+package com.cardinalblue.kraftshade.demo.ui.view
+
+import android.content.Context
+import android.widget.FrameLayout
+
+interface TraditionViewContent {
+    fun addContentTo(context: Context, container: FrameLayout)
+}
